requirements_image_processing.txt
# Requirements cho Image Processing và Computer Vision
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Optional: Cho advanced features
# ultralytics>=8.0.0  # YOLOv8 (nếu muốn dùng object detection model)
# torch>=2.0.0  # PyTorch (nếu muốn dùng deep learning models)

